requirements.txt
numpy>=1.26.4
imageio>=2.34.0
matplotlib>=3.7.3
scikit-learn>=1.3.1
joblib>=1.3.2
torch>=2.3.0+cu121
pandas>=2.0.2
opencv-python>=4.8.0.76
pillow>=10.2.0